update lastlog based on shell close, not on transport close also don't keep lastlog for exec commands, as unix does
